Added The Project and Dialogue Scriptis
Added the whole project with some added items (e.g. Panel for the dialogue UI and an item for DialogueManager) and two scripts, DialogueManager and DialogueType, where the first handles the dialogue and the second is the type that holds the name and the sentences of a dialogue.

Warning: When uploading the project, two files could not be copied (probably cause of Unity not wanting to share the project, cause one of the file names where named something with lock). The files where skipped, so if something misses might be them. Contact me to try to find a solution